web_service/requirements.txt
# pip install -e twitch_drops_watchdog
#requests==2.31.0
#google-api-python-client==2.102.0
google-cloud-firestore==2.12.0
#pytz==2023.3.post1
#jinja2==3.1.2